requirements-dev.txt
wand >= 0.6.10
fonttools
onnx
onnxscript
onnxconverter_common
numba
# https://github.com/richzhang/PerceptualSimilarity
lpips
# https://github.com/GallagherCommaJack/dctorch
dctorch
# https://github.com/facebookresearch/schedule_free
schedulefree
# https://github.com/toshas/torch-fidelity
torch-fidelity
